The most important change
BTW this is a funny joke, actually. You see, when I was writing this, I
took a mouse example from `usb-device` and with the help of usb hid spec
and Ben Eater videos, rewritten it to be a keyboard.

And so, it was not a mouse anymore, so I changed it:
```diff
-            .product("mouse")
+            .product("not mouse")
```

I don't know why, but I feel like "a" was missing this whole time idk