Sketch: Add physical layout interface
This commit feels comparatively very big...It is not *that* big -
`git diff --stat --staged` says 6 files changed, 179 insertions and
97 deletions. But still...

My current design idea for this library is to have +- 3 layers:
- Physical — represents where keys are physically located, how they are
  connected to the micro-controller, etc. It is sketched in this commit.
- Logical  — represents mapping of keys to their meaning. For example
  key#2 [on the default layer] is an English A,
  key#13 is a sticky shift,
  key#42 is a conditional layout switch,
  ...
  It is not yet implemented.
- Protocol — represents keyboard<->PC communications, in 99% of the
  cases it's USB. This is partially implemented, though I think I'll
  refine both interface and implementation in the future.

There is also a visual/topological representation that can allow
configuration interfaces to show roughly how keyboard looks. I don't
know how the interface should look and where it should be (I sketched it
as a submodule of physical layer), but I guess I'll figure this out with
time.

------------------------------------------------------------------------

I've also changed the example/testing binary. It previously continuously
printed the "Lorem ipsum [...]" text. It was ok as an example that shows
that I've made protocol layer work (at least a little bit). But it is
of course far from a real keyboard and I wanted an example that is
closer to the reality (which I hate, but still).

With this commit the example/testing binary (see `f103/` directory) is
now a 4 button keyboard! It is not much, but I'm still proud a bit :')

The setup is as follows:
- stm32f103c8 micro-controller aka bluepill
- 4 buttons
  - "Right" sides connected to the ground
  - "Left" sides connected to pins B12, B13, B14 and B15
- st-link (v2?) to upload binary to the micro-controller (& maybe debug
  but I haven't tried this in a long time)

Everything is very sketchy because I still don't have a soldering iron
and all the contacts are very bad (sometimes I need to hold wires in a
particular way just so everything works).

I have a picture of my sketchy setup, to get it run the following:
```shell
; git fetch origin "refs/notes/*:refs/notes/*"
; git notes --ref=files show commit > sketchy.jpg
```
Where `commit` is replaced with this commits hash and ; is shell prompt.

The picture is made on my pxl4a, then scaled down to 256x256 via Krita,
then drawn on top of via Krita and my mouse with #88ff82 color, and then
stored via git notes.

I've tried to compress it further, but it started loosing too much
details and wasn't vibing with me, so I stopped here.

And yes, I know that this is abuses git notes. 1) I don't care 2) I live
in another city, what can you do to me? 3) This has a great cursed vibe.

------------------------------------------------------------------------

As a side-note: I've noticed weird behavior with the example keyboard.
When holding multiple keys at the same time, only the last one is
repeated and when releasing the key that was last pressed, no keys are
repeated.

Example:
1. I press `A` (prints `AAAAA`...)
2. I press `B` (stops printing `A`s, prints `BBBBB`...)
3. I release `B` (prints nothing, even though `A` is still pressed)

I have no idea what causes this issue or how to fix it, so I just record
the fact that it happens here. That's all I can do at the moment.