Add support for modifiers to `UsbV1`
I've originally tried placing modifiers after the normal keys (which
would be more logical, if you ask me, considering their codes), but this
made holding `shift`+`a` print "aAAAAAA" because `a` was pressed before
shift from the perspective of the computer.

Also worth mentioning that when connecting to my phone, holding
`shift`+`a` with modifiers after the keys printed just a single "a".
Not sure why is that but probably this "only the last pressed key is
repeated" weird thing again?